+++ workers/api/test/env.d.ts
@@ -7,6 +7,7 @@ declare global {
   namespace Cloudflare {
     interface Env extends WorkerEnv {
       TEST_MIGRATIONS: D1Migration[];
+      TEST_RATE_LIMIT: number;
     }
   }
 }